requirements.txt
# Django Core
Django==4.2.7
psycopg2-binary==2.9.9

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Celery y Redis
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
celery[beat]==5.3.4

# Seguridad y Autenticación
django-axes==6.1.1
django-cors-headers==4.3.0
django-environ==0.11.2

# Almacenamiento y Archivos
django-storages==1.14.2
boto3==1.29.7
django-cleanup==8.0.0
python-magic==0.4.27

# Generación de PDFs y Documentos
WeasyPrint==60.1
python-docx==1.1.0

# Exportación de Datos
openpyxl==3.1.2
xlsxwriter==3.1.9

# Auditoría y Versionado
django-simple-history==3.4.0

# Búsqueda y Filtros
django-filter==23.3
django-contrib-postgres==0.0.1

# Validación y Utilidades
Pillow==10.1.0
python-dateutil==2.8.2

# Desarrollo y Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# ASGI Server (para producción)
gunicorn==21.2.0
uvicorn==0.24.0

# Monitoring y Logging
sentry-sdk==1.38.0
